### 2023-08-23 #603 @RWParsons
GLMMcosinor: Fit a cosinor model using the glmmTMB framework.
Submitting Author Name: Rex Parsons
Submitting Author Github Handle: @github_handle
Other Package Authors Github handles: (comma separated, delete if none) @oliverjayasinghe, @nicolemwhite
Repository:  https://github.com/RWParsons/GLMMcosinor
Version submitted: 0.1.0
Submission type: Stats
Badge grade: silver(select one)

Version accepted: TBD
---

```
Package: GLMMcosinor
Title: Fit a cosinor model using a generalised mixed modelling framework
Version: 0.1.0
    person("Rex", "Parsons", , "rex.parsons94@gmail.com", role = c("aut", "cre"),
    person("Oliver", "Jayasinghe", role = "aut"),
    person("Nicole", "White", role = "aut",
  )
Description: Fit the cosinor model using the glmmTMB framework, allowing
    for a wide range of link functions and capacity of heirachical
    structures.
    https://rwparsons.github.io/GLMMcosinor/
    assertthat,
    cowplot,
    ggforce,
    glmmTMB,
    lme4,
    scales,
    stats
    cosinor,
    covr,
    DT,
    flextable,
    ftExtra,
    vdiffr,
    withr
    knitr
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd",
    "srr::srr_stats_roclet"))
```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one appropriate box below):

- [x] A pre-submission inquiry has been approved in [issue#<601>](#<601>)

People analysing rhythmic/circular data - for example, circadian biologists.

This is the first implementation of a cosinor modelling package which can handle generalised models (link functions) in R. There are other packages in python but these are limited to count-data related families. Similarly, there are very limited other packages in R that can handle a hierarchical structure or have helpful plotting methods for the model objects. This package is based on the `{[cosinor](https://cran.r-project.org/web/packages/cosinor/index.html)}` R package but that is limited to linear models. A summary of existing software is given in a table in the README.

NA
