### 2023-08-25 555ee08
echo=FALSE for srr docs in vignette
