### 2023-08-25 4eb2cc4
Fix typos in docs and make plot legends consistent
