+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   model-specification: model-specification.html
   model-visualisations: model-visualisations.html
   multiple-components: multiple-components.html
-last_built: 2023-08-25T04:38Z
+last_built: 2023-08-25T07:14Z
 urls:
   reference: https://rwparsons.github.io/GLMMcosinor/reference
   article: https://rwparsons.github.io/GLMMcosinor/articles
