### 2023-08-29 811ec11
Reformat unnecessary single-length vectors
