+++ vignettes/GLMMcosinor.Rmd
@@ -151,8 +151,7 @@ acrophase ($\phi$), and MESOR ($M$).
 -   Acrophase ($\phi$) is the phase at which the maximal response
     occurs.
 
-In the single-component case, these parameters correspond to the
-following regression model:
+These parameters form the single-component cosinor model:
 
 $$Y(t) = M + Acos(\frac{2\pi t}{\tau} - \phi) + e(t)$$ where $e(t)$ is
 the error term.
@@ -164,19 +163,20 @@ $$Y(t) = M + \beta x + \gamma z + e(t)$$
 Where $x =cos(\frac{2\pi t}{τ})$, $z =sin(\frac{2\pi t}{τ})$,
 $\beta = A cos(\phi)$, $\gamma = A sin(\phi)$
 
-In our code, the `amp_acro()` function will modify the dataset supplied
-in the `cosinor.glmm()` call by creating two transformed columns of
-data: `rrr` and `sss` which take the role of $x$ and $z$ respectively
-from the equation above. 'rrr' and 'sss' are treated as variables, and
-the estimates of their coefficients are therefore $\hat\beta$ and
-$\hat\gamma$. These estimates are what is presented in as 'Raw
-Coefficients' in the output of the `cosinor.glmm()` function. Depending
-on the model specification, an estimate for MESOR ('(Intercept)') may be
-present and this does not need to be transformed. For this reason, the
-MESOR terms will be the same in both the raw and transformed coefficient
-tables in the output. But, to get the estimates for amplitude ($A$) and
-acrophase ($\phi$), the estimates for $\hat\beta$ and $\hat\gamma$ must
-be transformed as per the following equations:
+In the `GLMMcosinor` package, the `amp_acro()` function will modify the
+dataset supplied in the `cosinor.glmm()` call by creating two
+transformed columns of data: `rrr` and `sss` which take the role of $x$
+and $z$ respectively from the equation above. 'rrr' and 'sss' are
+treated as variables, and the estimates of their coefficients are
+therefore $\hat\beta$ and $\hat\gamma$. These estimates are what is
+presented in as 'Raw Coefficients' in the output of the `cosinor.glmm()`
+function. Depending on the model specification, an estimate for MESOR
+('(Intercept)') may be present and this does not need to be transformed.
+For this reason, the MESOR terms will be the same in both the raw and
+transformed coefficient tables in the output. But, to get the estimates
+for amplitude ($A$) and acrophase ($\phi$), the estimates for
+$\hat\beta$ and $\hat\gamma$ must be transformed as per the following
+equations:
 
 $$\hat\phi = \arctan(\frac{\hat\gamma}{\hat\beta}) $$
 
