+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   model-specification: model-specification.html
   model-visualisations: model-visualisations.html
   multiple-components: multiple-components.html
-last_built: 2023-08-29T12:39Z
+last_built: 2023-08-30T07:55Z
 urls:
   reference: https://rwparsons.github.io/GLMMcosinor/reference
   article: https://rwparsons.github.io/GLMMcosinor/articles
