### 2023-09-13 dc2dfb8
add ropensci badge to README and update NEWS.md
