### 2023-09-13 a13646e
Delete working_space.R
