+++ working_space.R
@@ -1,450 +1,0 @@
-# # working out space
-# comod = simulate_cosinor(1000,mesor = 50,amp = 7,acro = 1,beta.mesor = 0.1,beta.amp = 0.2, beta.acro = 0.3, family = "2_component")
-# m <- cosinor.glmm(Y ~ group+amp_acro(times, n_components = 2, group = "group", period = c(12, 12)), data = comod)
-# comod$pred <- predict(m$fit, type="response")
-#
-# ggplot() +
-#   geom_point(data=comod, aes(times, Y, col=as.factor(group)), alpha=0.1) +
-#   geom_line(data=comod, aes(times, pred, col=as.factor(group)))
-# # test when you put a browser() in at the simulate data stage with multicomponent (above)
-# data.frame(l1=lambda1, l2=lambda2, y=y, time=ttt) %>%
-#   mutate(lsum = l1+l2) %>%
-#   pivot_longer(!time) %>%
-#   ggplot(aes(time, value, col=name)) +
-#   geom_point()
-#
-#
-# #Example showing how cosinor.glmm works with a group of multiple levels
-# newgroup <-
-#   vitamind %>%
-#   filter(X==1) %>%
-#   mutate(X=2, time = time + 4)
-#
-# dat <- rbind(vitamind, newgroup) %>%
-#   mutate(X = as.factor(X))
-#
-# dat %>%
-#   ggplot(aes(time, Y, col=X)) + geom_point()
-#
-# m <- cosinor.glmm(Y ~ group+amp_acro(time, n_components = 2, group = "X", period = c(12,10)), data = dat)
-#
-# # Examples for meeting: 15/12/2022
-# # Evidence that the multi-component estimation works. In this example, amp2 = amp1 + 3, acr2 = acr1-1.
-# comod = simulate_cosinor(100,mesor = 1,amp = 2,acro = 3,beta.mesor = 0.5,beta.amp = 0.2, beta.acro = 0.3, family = "2_component")
-# cosinor.glmm(Y ~ group+amp_acro(times, n_components = 2, group = "group", period = c(12, 8)), data = comod)
-# #
-# # You can now have different groups assigned to different components. For example:
-# vitamind$Z = rbinom(length(vitamind$X),1,prob = 0.5)
-# cosinor.glmm(Y~ X + amp_acro(time, n_components = 3, group = c("X",NA,"Z"), period = c(12,10,8)),data = vitamind)
-#
-# dat2 = dat
-# dat2$Z = (rbinom(length(dat2$X),5,prob = 0.5))
-# cosinor.glmm(Y~ X + Z + amp_acro(time, n_components = 3, group = c("X",NA,"Z"), period = c(12,10,8)),data = dat2)
-#
-#
-# #Meeting for 19/12/2022
-# #Added comments to the cosinor.glmm script for everything except the update_covnames function
-#   #Do we need the update_covnames? Currently, it is not being used in the script.
-# #Added the ability to have groups be factors with more than 2 levels.Example:
-# cosinor.glmm(Y ~ group+amp_acro(time, n_components = 2, group = "X", period = c(12,10)), data = dat)
-# #Added some error messages. For example, if n_components is not an integer, or it is less than 1,
-# #an error will appear with a relevant message. Examples:
-# cosinor.glmm(Y ~ group+amp_acro(time, n_components = 2.1, group = "X", period = c(12,10)), data = dat)
-# cosinor.glmm(Y ~ group+amp_acro(time, n_components = 0, group = "X", period = c(12,10)), data = dat)
-# #Changed some variable names to be more descriptive and informative.
-# #If for whatever reason the user names their grouping variables 'rrr' or 'sss', there will be
-# #problems when getting the terms for the formula. Should we add a simple error message if the user
-# #uses these names for their groups? The code in question is this:
[…]
