### 2023-09-24 b4f3d8f
autoplot() now works for mixed models
