+++ R/autoplot.R
@@ -67,6 +67,7 @@ autoplot.cosinor.glmm <- function(object,
     msg = "'object' must be of class 'cosinor.glmm'"
   )
 
+
   validate_ci_level(ci_level)
 
   if (!missing(x_str)) {
@@ -128,6 +129,23 @@ autoplot.cosinor.glmm <- function(object,
     )
   }
 
+
+  #if there are multiple random effects groups, the time vector must be scaled
+  #appropriately. Otherwise, the number of simulated datapoints may be
+  #extremely large. For example, if pred.length.out = 400, and there are two
+  #random effects groups with 30 subjects each, then the final dataframe will
+  #be 400*30*30 = 360000 rows. Instead, this code restricts the number of data
+  #points to 10000. In future versions, this could be set to a variable that
+  #the user can input
+  ranef_scale <- 1
+  for (i in object$ranef_groups) {
+  ranef_scale <- ranef_scale*length(levels(object$newdata[[i]]))
+  }
+
+  if(ranef_scale*pred.length.out >= 10000) {
+    pred.length.out <- 10000 %/% (ranef_scale)
+  }
+
   # generate the time values for the x-axis
   if (!missing(xlims)) {
     # with multiple periods, largest is used for timeax simulation
@@ -152,7 +170,6 @@ autoplot.cosinor.glmm <- function(object,
       ref_level <- unlist(x$group_stats[j])[[1]]
       newdata[, j] <- factor(ref_level)
     }
-
     # process the data. This step mimics the first step of a cosinor.glmm() call
     newdata <- update_formula_and_data(
       # pass new dataset that's being used for prediction in this function
@@ -188,6 +205,53 @@ autoplot.cosinor.glmm <- function(object,
         i = i + 1
       }
     }
+
+    ##
+    #ranef_part <- lapply(lme4::findbars(object$formula), deparse1)
+    #ranef_part_group <- gsub(".*\\|\\s*(.*)", "\\1", ranef_part)
+    ##
+   # ## Here, 'x_str' is taken to be the string corresponding to the subjects
+   # if (!is.null(ranef_part)) {
+   #   if(is.null(x_str)) {
+   #     x_str <- ranef_part_group[1]
+   #   }
+    #
+   #   ###
+   #   mixed_plot <- TRUE
+   #   mixed_effects_vec <- NULL
+   #   replicated_dfs <- NULL
+
+    if(any(!is.na(object$ranef_groups))) {
+   replicated_dfs_total <- NULL
+    data_ranef <- newdata
+   for (i in x$ranef_groups) {
+     replicated_dfs <- NULL
+     subjects <- as.factor(levels(x$newdata[[i]]))
+     for (j in subjects){
+       appendvec <- data_ranef
+       appendvec[[i]] <- j
+       replicated_dfs[[j]] <- appendvec
+     }
+     replicated_dfs_total <- do.call(rbind,replicated_dfs)
+     data_ranef <- replicated_dfs_total
+
+   }
+   #combined_df <- do.call(rbind, replicated_dfs_total)
+   newdata <- data_ranef
+    }
+   #   subjects <- as.factor(levels(object$newdata[[x_str]]))
+   #   for (m in subjects) {
+  #
+   #     appendvec <- newdata
+   #     appendvec$x_str <- m
+   #     replicated_dfs[[m]] <- appendvec
+   #   }
+   #   combined_df <- do.call(rbind, replicated_dfs)
+   #   ###
+   #   newdata <- combined_df
+   # }
+    ##
+
     newdata
   }
 
@@ -198,7 +262,6 @@ autoplot.cosinor.glmm <- function(object,
 
   # get the response data from the cosinor.glmm object
   y_name <- object$response_var
-
   # get the predicted response values using the predict.cosinor.glmm() function
[…]
