+++ vignettes/mixed-models.Rmd
@@ -100,6 +100,12 @@ This works by replacing the amp_acro1 with the relevant cosinor components when 
 mixed_mod$formula
 ```
 
+The mixed model can also be plotted using `autoplot`, but some of the plotting features that are available for fixed-effects models may not be available for mixed-effect model.
+
+```{r}
+autoplot(mixed_mod)
+```
+
 The summary of the model shows that the input means for MESOR, amplitude and acrophase are similar to what we specified in the simulation (0, 3, and 1.5, respectively).
 
 ```{r}
