### 2023-09-24 8afb22d
fix grammatical error in mixed-model vignette
