### 2023-09-26 6e271ff
superimpose data on mixed-model plot in mixed-models vignette
