+++ vignettes/mixed-models.Rmd
@@ -103,7 +103,7 @@ mixed_mod$formula
 The mixed model can also be plotted using `autoplot`, but some of the plotting features that are available for fixed-effects models may not be available for mixed-effect models.
 
 ```{r}
-autoplot(mixed_mod)
+autoplot(mixed_mod, superimpose.data = TRUE)
 ```
 
 The summary of the model shows that the input means for MESOR, amplitude and acrophase are similar to what we specified in the simulation (0, 3, and 1.5, respectively).
