### 2023-09-26 23a257d
updates from meeting
