+++ R/data_utils.R
@@ -91,7 +91,6 @@ update_formula_and_data <- function(data,
     e$.quietly <- quietly
     e$.amp_acro_ind <- amp_acro_ind
     e$.data_prefix <- data_prefix
-
     #
     ranef_part <- lapply(lme4::findbars(formula), deparse1)
     ranef_part_group <- gsub(".*\\|\\s*(.*)", "\\1", ranef_part)
