### 2023-09-28 2d5db43
changes to how autoplot generates mixed model plots
