+++ R/autoplot.R
@@ -61,6 +61,7 @@ autoplot.cosinor.glmm <- function(object,
                                   superimpose.data = FALSE,
                                   data_opacity = 0.3,
                                   predict.ribbon = TRUE,
+                                  ranef_plot = NULL,
                                   ...) {
   # Validating user inputs
   assertthat::assert_that(inherits(object, "cosinor.glmm"),
@@ -137,14 +138,14 @@ autoplot.cosinor.glmm <- function(object,
   #be 400*30*30 = 360000 rows. Instead, this code restricts the number of data
   #points to 10000. In future versions, this could be set to a variable that
   #the user can input
-  ranef_scale <- 1
-  for (i in object$ranef_groups) {
-  ranef_scale <- ranef_scale*length(levels(object$newdata[[i]]))
-  }
-
-  if(ranef_scale*pred.length.out >= 10000) {
-    pred.length.out <- 10000 %/% (ranef_scale)
-  }
+  #ranef_scale <- 1
+  #for (i in object$ranef_groups) {
+  #ranef_scale <- ranef_scale*length(levels(object$newdata[[i]]))
+  #}
+#
+  #if(ranef_scale*pred.length.out >= 10000) {
+  #  pred.length.out <- 10000 %/% (ranef_scale)
+  #}
 
   # generate the time values for the x-axis
   if (!missing(xlims)) {
@@ -206,56 +207,60 @@ autoplot.cosinor.glmm <- function(object,
       }
     }
 
-    ##
-    #ranef_part <- lapply(lme4::findbars(object$formula), deparse1)
-    #ranef_part_group <- gsub(".*\\|\\s*(.*)", "\\1", ranef_part)
-    ##
-   # ## Here, 'x_str' is taken to be the string corresponding to the subjects
-   # if (!is.null(ranef_part)) {
-   #   if(is.null(x_str)) {
-   #     x_str <- ranef_part_group[1]
-   #   }
-    #
-   #   ###
-   #   mixed_plot <- TRUE
-   #   mixed_effects_vec <- NULL
-   #   replicated_dfs <- NULL
 
+    #adding the random effect variables in the dataframe
     if(any(!is.na(object$ranef_groups))) {
    replicated_dfs_total <- NULL
     data_ranef <- newdata
+
+    unique_counts <- sapply(object$newdata[ranef_plot],
+                            function(col) length(unique(col)))
+
+    # Identify the column with the maximum number of unique values
+    max_unique_col <- names(unique_counts)[which.max(unique_counts)]
+    # Create a new vector excluding the max_unique_col
+    new_ranef_groups <- ranef_plot[ranef_plot != max_unique_col]
+
    for (i in x$ranef_groups) {
      replicated_dfs <- NULL
      subjects <- as.factor(levels(x$newdata[[i]]))
-     for (j in subjects){
+
+     #
+     if (!is.null(ranef_plot) && i %in% ranef_plot && i == max_unique_col) {
+       for (j in subjects){
+         appendvec <- data_ranef
+         appendvec[[i]] <- j
+         for (k in new_ranef_groups) {
+           appendvec[[k]] <- unique(
+             object$newdata[object$newdata$subject == 1, k])
+         }
+
+         replicated_dfs[[j]] <- appendvec
+       }
+     } else {
        appendvec <- data_ranef
-       appendvec[[i]] <- j
-       replicated_dfs[[j]] <- appendvec
+       appendvec[[i]] <- rep(NA, length(ncol(appendvec)))
+       replicated_dfs[[i]] <- appendvec
      }
+
+
+
      replicated_dfs_total <- do.call(rbind,replicated_dfs)
      data_ranef <- replicated_dfs_total
-
    }
    #combined_df <- do.call(rbind, replicated_dfs_total)
    newdata <- data_ranef
     }
[…]
