### 2023-09-28 3337dd5
fix bugs in autoplot()
