### 2023-10-03 #603 @sachsmc
## Package Review

-   ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   ☒ **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README
-   ☒ **Installation instructions:** for the development version of package and any non-standard dependencies in README
-   ☒ **Vignette(s):** demonstrating major functionality that runs successfully locally
-   ☒ **Function Documentation:** for all exported functions
-   ☒ **Examples:** (that run successfully locally) for all exported functions
-   ☒ **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

#### Functionality

-   ☒ **Installation:** Installation succeeds as documented.
-   ☒ **Functionality:** Any functional claims of the software been confirmed.
-   ☒ **Performance:** Any performance claims of the software been confirmed.
-   ☒ **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
-   ☒ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.

Estimated hours spent reviewing: 4

-   ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

### Review Comments

-   The package is quite high quality and I am impressed by the breadth of features available and their implementations. I think it will be incredibly useful to the research community working with rhythmic data. I will recommend this to anyone who asks how to analyze rhythmic data.
-   The package docs are quite impressive to someone who is familiar with the cosinor model, but maybe it is not so accessible to someone who is not. To remedy this, I think the vignettes and possibly the readme could use a more thorough description of the motivation for using the cosinor model. Maybe in the context of the vitamind example dataset you could give a short explanation of the biological rationale for the model, what the potential research questions that could be answered, and how the data were (hypothetically) collected.
-   Feature suggestion: predict function could be enhanced to also obtain predictions of time-independent parameters given a vector of covariates, e.g., the amplitude, acrophase, or MESOR.
-   The title in DESCRIPTION will need to be in title case if you plan to submit to CRAN.
-   Inconsistent use of snake case and dot case in the function arguments of: autoplot, simulate\_cosinor
-   The vitamind example dataset description is completely generic (e.g., “Y is an outcome variable”). You could consider adding some fake details about the measurements, or instead changing the name of the dataset to something equally generic, like example\_data or similar.
