### 2023-10-04 #603 @RWParsons
Thank you very much @sachsmc for the positive feedback and helpful comments!
We will work to address them over the coming couple weeks but I'd just like to clarify how you would picture the use of the proposed feature for the predict function?

Are you suggesting that the user pass the inputs and the output would be a visualised rhythm?

say something like:

```
predict(MESOR = 5, amplitude= 2, acrophase = pi, period = 12)
```

And it would spit out something similar to `simulate_cosinor()` but with no noise term? I think I'm struggling to distinguish how this would be very different from `simulate_cosinor()` (without an error term) so it likely that I'm not understanding how you intend for it to work.

Thanks!
