### 2023-10-04 #603 @sachsmc
No, not a visualized rhythm, but more like a summary statistic of the rhythm for a vector of covariate values. So the main arguments would be the newdata, but then you could add a type = "amplitude", or "mesor", for example, which would give the predicted parameter for the subpopulation with those covariate values. For simple models, those parameters can be obtained directly from the print method, but this predict approach would be useful for complex models with continuous covariates, for example. I had the predict methods from the survival package in mind when I made the suggestion.

Come to think of it, is it true that it is not possible to specify a model such that the amplitude and acrophase vary continuously in the covariate? Is there a need for such a model or would that be too biologically implausible?

Anyway these are just minor suggestions, so if this is to complex or there is no need for such things, feel free to ignore.
