### 2023-10-04 #603 @RWParsons
Oh right!

You're right, it's currently not possible to specify a model such that the amplitude/acrophase can vary over time. It can only vary in relation the grouping levels but that's provided to the user when using `print(model)`.

I've given some thought to the type of model that you're suggesting before though. circacompare allows the user to specify a decay term on amplitude or acrophase so this feature may be useful there, but it's only possible because it uses nonlinear regression instead of the cosinor model. Otherwise, I think a cosinor model would need to be adjusted to have an interaction term between time (or possibly a nonlinear spline on time) with the cosinor components (rrr and sss). This could sort of allow the confidence ellipse of a rhythm "move" over time and be could be evaluated at any given time-point. (I was thinking to do this for evaluating how fast an animal adjusts to a phase shift but never pursued it.)

This feature might be out of scope for this initial version but I'll keep in mind for perhaps a future development of the package. I think it'd probably require some adjustments to `amp_acro()` to allow the user to specify an interaction term with time and that'd have some flow on effects to the summary/print/autplot/polar_plot methods.
