### 2023-10-04 ropensci/GLMMcosinor#2 @RWParsons
rOpenSci reviewer feedback from @sachsmc
Feedback from @sachsmc here <https://github.com/ropensci/software-review/issues/603>

**Review Comments:**

- [x] The package docs are quite impressive to someone who is familiar with the cosinor model, but maybe it is not so accessible to someone who is not. To remedy this, I think the vignettes and possibly the readme could use a more thorough description of the motivation for using the cosinor model. Maybe in the context of the vitamind example dataset you could give a short explanation of the biological rationale for the model, what the potential research questions that could be answered, and how the data were (hypothetically) collected.
- [x]  ~~Feature suggestion: predict function could be enhanced to also obtain predictions of time-independent parameters given a vector of covariates, e.g., the amplitude, acrophase, or MESOR.~~
- [x]  The title in DESCRIPTION will need to be in title case if you plan to submit to CRAN.
- [x] Inconsistent use of snake case and dot case in the function arguments of: autoplot, simulate_cosinor
- [x] The vitamind example dataset description is completely generic (e.g., “Y is an outcome variable”). You could consider adding some fake details about the measurements, or instead changing the name of the dataset to something equally generic, like example_data or similar.
