### 2023-10-04 RWParsons/GLMMcosinor#2 @RWParsons
Point 2 (Feature suggestion) has been noted for a possible future work but will not be addressed as part of this submission/review. See <https://github.com/ropensci/software-review/issues/603#issuecomment-1747723048> for a discussion/response to this comment.
