+++ R/test_cosinor.R
@@ -7,7 +7,7 @@
 #' continuous covariates.
 #'
 #'
-#' @param x An \code{cosinor.glmm} object.
+#' @param x An \code{cglmm} object.
 #' @param x_str A \code{character}. The name of the grouping variable within
 #' which differences in the selected cosinor characteristic (amplitude or
 #' acrophase) will be tested.
@@ -42,7 +42,7 @@
 #'   period = c(10, 12),
 #'   beta.group = TRUE
 #' )
-#' mod_2_component <- cosinor.glmm(
+#' mod_2_component <- cglmm(
 #'   Y ~ group + amp_acro(times,
 #'     n_components = 2, group = "group",
 #'     period = c(10, 12)
@@ -69,8 +69,8 @@ test_cosinor_components <- function(x,
   validate_ci_level(ci_level)
 
   assertthat::assert_that(
-    inherits(x, "cosinor.glmm"),
-    msg = "'x' must be of class 'cosinor.glmm'"
+    inherits(x, "cglmm"),
+    msg = "'x' must be of class 'cglmm'"
   )
 
   stopifnot(is.character(x_str))
@@ -118,7 +118,7 @@ test_cosinor_components <- function(x,
 #' group with covariates equal to 0. This may not be the desired result for
 #' continuous covariates.
 #'
-#' @param x An \code{cosinor.glmm} object.
+#' @param x An \code{cglmm} object.
 #' @param x_str A \code{character}. The name of the grouping variable within
 #' which differences in the selected cosinor characteristic (amplitude or
 #' acrophase) will be tested.
@@ -156,7 +156,7 @@ test_cosinor_components <- function(x,
 #'   period = c(10, 12),
 #'   beta.group = TRUE
 #' )
-#' mod_2_component <- cosinor.glmm(
+#' mod_2_component <- cglmm(
 #'   Y ~ group + amp_acro(times,
 #'     n_components = 2, group = "group",
 #'     period = c(10, 12)
@@ -182,8 +182,8 @@ test_cosinor_levels <- function(x,
   validate_ci_level(ci_level)
 
   assertthat::assert_that(
-    inherits(x, "cosinor.glmm"),
-    msg = "'x' must be of class 'cosinor.glmm'"
+    inherits(x, "cglmm"),
+    msg = "'x' must be of class 'cglmm'"
   )
 
   stopifnot(is.character(x_str))
@@ -249,7 +249,7 @@ test_cosinor_levels <- function(x,
 #' continuous covariates.
 #'
 #'
-#' @param x An \code{cosinor.glmm} object.
+#' @param x An \code{cglmm} object.
 #' @param x_str A \code{character}. The name of the grouping variable within
 #' which differences in the selected cosinor characteristic (amplitude or
 #' acrophase) will be tested.
@@ -277,7 +277,7 @@ test_cosinor_levels <- function(x,
 #'
 #' @return Returns a \code{test_cosinor} object.
 #' @examples
-#' mod_grouped <- cosinor.glmm(Y ~ X + amp_acro(time,
+#' mod_grouped <- cglmm(Y ~ X + amp_acro(time,
 #'   group = "X",
 #'   n_components = 1,
 #'   period = 12
@@ -298,7 +298,7 @@ test_cosinor_levels <- function(x,
 #'   period = c(10, 12)
 #' )
 #'
-#' mod_2_component <- cosinor.glmm(
+#' mod_2_component <- cglmm(
 #'   Y ~ group + amp_acro(times,
 #'     n_components = 2, group = "group",
 #'     period = c(10, 12)
@@ -321,7 +321,7 @@ test_cosinor_levels <- function(x,
   # TODO handle the text labels for the factor variable used for grouping - i.e
   # using data made with : sim_data %>% mutate(group = ifelse(group == 1,
[…]
