### 2023-10-05 234aca3
(addressing rOpenSci comments #2) name changes
