+++ tests/testthat/test-test_cosinor.R
@@ -167,8 +167,10 @@ test_that("script works and warnings are displayed appropriately", {
   }
   expect_error(
     f(),
-    regex = paste("'comparison_A' and 'comparison_B' must be numbers",
-                  "corresponding to a component in the model"), fixed = TRUE
+    regex = paste(
+      "'comparison_A' and 'comparison_B' must be numbers",
+      "corresponding to a component in the model"
+    ), fixed = TRUE
   )
 
 
@@ -190,8 +192,10 @@ test_that("script works and warnings are displayed appropriately", {
   }
   expect_error(
     f(),
-    regex = paste("'component_index' must be supplied and it must be a",
-                  "number corresponding to a component in the model"),
+    regex = paste(
+      "'component_index' must be supplied and it must be a",
+      "number corresponding to a component in the model"
+    ),
     fixed = TRUE
   )
 
@@ -215,8 +219,10 @@ test_that("script works and warnings are displayed appropriately", {
   }
   expect_error(
     f(),
-    regex = paste("'level_index' must be supplied and it must be a number",
-                  "corresponding to a level in the model"), fixed = TRUE
+    regex = paste(
+      "'level_index' must be supplied and it must be a number",
+      "corresponding to a level in the model"
+    ), fixed = TRUE
   )
 
 
