+++ tests/testthat/test-amp-acro.R
@@ -172,8 +172,10 @@ test_that("bad inputs return useful errors", {
     )
   }
   expect_error(f(),
-    regexp = paste("Grouping variable in amp_acro() must be of",
-                   "length 1 or the same as n_components"),
+    regexp = paste(
+      "Grouping variable in amp_acro() must be of",
+      "length 1 or the same as n_components"
+    ),
     fixed = TRUE
   )
 
@@ -191,8 +193,10 @@ test_that("bad inputs return useful errors", {
     )
   }
   expect_error(f(),
-    regexp = paste("period value(s) in amp_acro() must be of",
-                   "length 1 or the same as n_components"),
+    regexp = paste(
+      "period value(s) in amp_acro() must be of",
+      "length 1 or the same as n_components"
+    ),
     fixed = TRUE
   )
 
