+++ R/data_utils.R
@@ -95,7 +95,7 @@ update_formula_and_data <- function(data,
     ranef_part <- lapply(lme4::findbars(formula), deparse1)
     ranef_part_group <- gsub(".*\\|\\s*(.*)", "\\1", ranef_part)
     #
-    #e$subject <- ranef_part_group
+    # e$subject <- ranef_part_group
     updated_df_and_formula <- eval(e) # evaluate amp_acro call
     c(updated_df_and_formula, list(
       Terms = Terms, family = family,
