### 2023-10-05 32f9bc1
(#2) update DESCRIPTION to use title case for Title
