+++ man/summary.cglmm.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/summary.cosinor.glmm.R
+% Please edit documentation in R/summary.R
 \name{summary.cglmm}
 \alias{summary.cglmm}
 \title{Summarize a cosinor model
