+++ man/cglmm.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/cosinor.glmm.R
+% Please edit documentation in R/cglmm.R
 \name{cglmm}
 \alias{cglmm}
 \title{Fit cosinor model with \code{{glmmTMB}}}
