+++ man/predict.cglmm.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/predict.cosinor.glmm.R
+% Please edit documentation in R/predict.R
 \name{predict.cglmm}
 \alias{predict.cglmm}
 \title{Predict from a cosinor model}
