+++ man/print.cglmmSummary.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/summary.cosinor.glmm.R
+% Please edit documentation in R/summary.R
 \name{print.cglmmSummary}
 \alias{print.cglmmSummary}
 \title{Print the summary of a cosinor model}
