+++ man/print.cglmmTest.Rd
@@ -33,8 +33,8 @@ data_2_component <- simulate_cosinor(
 )
 mod_2_component <- cglmm(
   Y ~ group + amp_acro(times,
-                       n_components = 2, group = "group",
-                       period = c(10, 12)
+    n_components = 2, group = "group",
+    period = c(10, 12)
   ),
   data = data_2_component
 )
