+++ DESCRIPTION
@@ -46,3 +46,5 @@ LazyData: true
 Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd",
     "srr::srr_stats_roclet"))
 RoxygenNote: 7.2.3
+Depends: 
+    R (>= 2.10)
