+++ R/amp_acro.R
@@ -33,7 +33,7 @@
 #' @examples
 #' # Single component cosinor model
 #' cglmm(
-#'   Y ~ amp_acro(time_col = time, group = "X", period = 12),
+#'   vit_d ~ amp_acro(time_col = time, group = "X", period = 12),
 #'   data = vitamind
 #' )
 #'
