+++ R/print.R
@@ -12,7 +12,7 @@
 #' @examples
 #' # Single component cosinor model
 #' cglmm(
-#'   Y ~ amp_acro(time_col = time, group = "X", period = 12),
+#'   vit_d ~ amp_acro(time_col = time, group = "X", period = 12),
 #'   data = vitamind
 #' )
 #' @export
