+++ R/data.R
@@ -1,10 +1,13 @@
-#' Vitamin D
+#' Vitamin D dataset for cosinor modelling examples.
 #'
-#' Simulated data set to illustrate the cosinor model. \code{Y}
-#' is an outcome variable that varies of time \code{time} according
-#' to a cosine curve. The binary covariate \code{X} is associated with the
-#' mean and amplitude of the cosine curve.
+#' Simulated data set to illustrate the cosinor model. The \code{vit_d} column
+#' contains the blood vitamin D levels which vary over time (\code{time}).
+#' The rhythm of the vitamind D fluctuations follows a cosine function and can
+#' be modelled with a cosinor model. The \code{X} column is a binary covariate
+#' representing two groups of patients and is associated with the
+#' characteristics of the rhythm. The rhythm has a period of about 12 hours.
 #'
-#' @format A data frame with 3 variables: \code{X}, \code{Y}, \code{time}.
+#' @format A \code{data.frame} with 3 variables: \code{vit_d}, \code{time}, and
+#' \code{X}.
 
 "vitamind"
