+++ R/data_utils.R
@@ -33,7 +33,7 @@
 #' # without yet fitting the model
 #' data_and_formula <- update_formula_and_data(
 #'   data = vitamind2,
-#'   formula = Y ~ X + amp_acro(time,
+#'   formula = vit_d ~ X + amp_acro(time,
 #'     group = "X",
 #'     period = 12
 #'   )

[…]
