### 2023-10-06 09a41ad
fix up spelling and create WORDLIST
