+++ DESCRIPTION
@@ -10,7 +10,7 @@ Authors@R: c(
            comment = c(ORCID = "0000-0002-9292-0773"))
   )
 Description: Fit the cosinor model using the glmmTMB framework, allowing
-    for a wide range of link functions and capacity of heirachical
+    for a wide range of link functions and capacity of hierarchical
     structures.
 License: GPL (>= 3)
 URL: https://github.com/RWParsons/GLMMcosinor,
