### 2023-10-07 4100aae
add tests for r'anef_plot' argument in autoplot()
