### 2023-10-07 4a73eb1
remove redundant code in autoplot()
