+++ vignettes/GLMMcosinor.Rmd
@@ -119,18 +119,18 @@ dat
 
 ## `cglmm()`
 
-`cglmm()` wrangles the data appropriately to fit the cosinor
-model given the formula specified by the user. It provides estimates of
-amplitude, acrophase, and MESOR (Midline Statistic Of Rhythm).
+`cglmm()` wrangles the data appropriately to fit the cosinor model given
+the formula specified by the user. It provides estimates of amplitude,
+acrophase, and MESOR (Midline Statistic Of Rhythm).
 
-The formula argument for `cglmm()` is specified using the
-`{lme4}` style (for details see `vignette("lmer", package = "lme4")`).
-The only difference is that it allows for use of an `amp_acro()` call
-within the formula that is used to identify the circadian components and
-relevant variables in the `data.frame`. Any other combination of
-covariates can also be included in the formula as well as random effects
-and zero-inflation (`ziformula`) and dispersion (`dispformula`)
-formulae. For detailed examples of how to specify models, see the
+The formula argument for `cglmm()` is specified using the `{lme4}` style
+(for details see `vignette("lmer", package = "lme4")`). The only
+difference is that it allows for use of an `amp_acro()` call within the
+formula that is used to identify the circadian components and relevant
+variables in the `data.frame`. Any other combination of covariates can
+also be included in the formula as well as random effects and
+zero-inflation (`ziformula`) and dispersion (`dispformula`) formulae.
+For detailed examples of how to specify models, see the
 [mixed-models](https://rwparsons.github.io/GLMMcosinor/articles/mixed-models.html),
 [model-specification](https://rwparsons.github.io/GLMMcosinor/articles/model-specification.html)
 and
@@ -164,19 +164,18 @@ Where $x =cos(\frac{2\pi t}{τ})$, $z =sin(\frac{2\pi t}{τ})$,
 $\beta = A cos(\phi)$, $\gamma = A sin(\phi)$
 
 In the `GLMMcosinor` package, the `amp_acro()` function will modify the
-dataset supplied in the `cglmm()` call by creating two
-transformed columns of data: `rrr` and `sss` which take the role of $x$
-and $z$ respectively from the equation above. 'rrr' and 'sss' are
-treated as variables, and the estimates of their coefficients are
-therefore $\hat\beta$ and $\hat\gamma$. These estimates are what is
-presented in as 'Raw Coefficients' in the output of the `cglmm()`
-function. Depending on the model specification, an estimate for MESOR
-('(Intercept)') may be present and this does not need to be transformed.
-For this reason, the MESOR terms will be the same in both the raw and
-transformed coefficient tables in the output. But, to get the estimates
-for amplitude ($A$) and acrophase ($\phi$), the estimates for
-$\hat\beta$ and $\hat\gamma$ must be transformed as per the following
-equations:
+dataset supplied in the `cglmm()` call by creating two transformed
+columns of data: `rrr` and `sss` which take the role of $x$ and $z$
+respectively from the equation above. 'rrr' and 'sss' are treated as
+variables, and the estimates of their coefficients are therefore
+$\hat\beta$ and $\hat\gamma$. These estimates are what is presented in
+as 'Raw Coefficients' in the output of the `cglmm()` function. Depending
+on the model specification, an estimate for MESOR ('(Intercept)') may be
+present and this does not need to be transformed. For this reason, the
+MESOR terms will be the same in both the raw and transformed coefficient
+tables in the output. But, to get the estimates for amplitude ($A$) and
+acrophase ($\phi$), the estimates for $\hat\beta$ and $\hat\gamma$ must
+be transformed as per the following equations:
[…]
