### 2023-10-07 3140972
increase coverage of test-autoplot()
