### 2023-10-07 c128683
increase coverage further for autoplot()
