### 2023-10-07 ace19fa
increase coverage of test script for summary() function
