### 2023-10-09 ff6c8b8
make tests for autoplot() and polar_plot() more specific
