+++ tests/testthat/test-autoplot.R
@@ -4,8 +4,7 @@
 #' @srrstats {G5.2a}
 #' @srrstats {G5.2b}
 
-
-test_that("autoplot input checks work", {
+test_that("autoplot works with simple inputs", {
   data(vitamind)
   object <- cglmm(
     vit_d ~ amp_acro(time,
@@ -15,14 +14,20 @@ test_that("autoplot input checks work", {
     data = vitamind
   )
 
-  # testing various inputs
-  expect_no_error(autoplot(object, superimpose.data = TRUE))
-  expect_no_error(autoplot(object, predict.ribbon = TRUE))
+  # testing autoplot with a simple model
+  vdiffr::expect_doppelganger("plot with superimposed data",
+                              autoplot(object, superimpose.data = TRUE))
+  vdiffr::expect_doppelganger("plot with predict ribbon enabled",
+                              autoplot(object, predict.ribbon = TRUE))
 })
 
 
 test_that("autoplot produces error messages", {
   data(vitamind)
+
+  #Testing error messages
+
+  # test 'x_str' argument
   object <- cglmm(
     vit_d ~ 1 + amp_acro(time,
       group = "X",
@@ -31,15 +36,6 @@ test_that("autoplot produces error messages", {
     data = vitamind
   )
 
-  # Test 1
-  data(vitamind)
-  object <- cglmm(
-    vit_d ~ 1 + amp_acro(time,
-      group = "X",
-      period = 12
-    ),
-    data = vitamind
-  )
   f <- function() {
     autoplot(object, x_str = 10)
   }
@@ -52,32 +48,36 @@ test_that("autoplot produces error messages", {
     ), fixed = TRUE
   )
 
-  # Test 2
-  data(vitamind)
-  vitamind$Z <- rbinom(length(vitamind$X), 3, prob = 0.5)
-  object <- cglmm(
-    vit_d ~ X + amp_acro(time,
-      n_components = 3,
-      group = c("Z", NA, "X"),
-      period = c(12, 10, 8)
-    ),
-    data = vitamind
+  # test plots with different groups on different components
+  withr::with_seed(
+    50,{
+
+      vitamind$Z <- rbinom(length(vitamind$X), 3, prob = 0.5)
+      object <- cglmm(
+        vit_d ~ X + amp_acro(time,
+                             n_components = 3,
+                             group = c("Z", NA, "X"),
+                             period = c(12, 10, 8)
+        ),
+        data = vitamind
+      )
+
+      f <- function() {
+        autoplot(object, x_str = c("X", "Z"), pred.length.out = 200)
+      }
+      vdiffr::expect_doppelganger("check plots with multiple groups",f())
+
+      f <- function() {
+        autoplot(object)
+      }
+
+      vdiffr::expect_doppelganger("check a simple plot command", f())
+
+    }
   )
 
-  f <- function() {
-    autoplot(object, x_str = c("X", "Z"), pred.length.out = 200)
-  }
 
-  expect_no_error(f())
-
-  f <- function() {
-    autoplot(object)
-  }
-
-  # Test 3
-  expect_no_error(f())
-
-  # Test 4
+  # test 'type' argument
   data(vitamind)
   object <- cglmm(
     vit_d ~ 1 + amp_acro(time,
@@ -98,8 +98,7 @@ test_that("autoplot produces error messages", {
     ), fixed = TRUE
   )
 
-  # Test 5
-  data(vitamind)
+  # test 'xlims' argument
   object <- cglmm(
     vit_d ~ 1 + amp_acro(time,
       group = "X",
@@ -121,8 +120,7 @@ test_that("autoplot produces error messages", {
     fixed = TRUE
   )
 
-  # Test 6
[…]
