+++ vignettes/model-specification.Rmd
@@ -78,18 +78,18 @@ withr::with_seed(
 
 ## `cglmm()`
 
-`cglmm()` wrangles the data appropriately to fit the cosinor
-model given the formula specified by the user. It provides estimates of
-amplitude, acrophase, and MESOR (Midline Statistic Of Rhythm).
+`cglmm()` wrangles the data appropriately to fit the cosinor model given
+the formula specified by the user. It provides estimates of amplitude,
+acrophase, and MESOR (Midline Statistic Of Rhythm).
 
-The formula argument for `cglmm()` is specified using the
-`{lme4}` style (for details see `vignette("lmer", package = "lme4")`).
-The only difference is that it allows for use of an `amp_acro()` call
-within the formula that is used to identify the circadian components and
-relevant variables in the `data.frame`. Any other combination of
-covariates can also be included in the formula as well as random effects
-and zero-inflation (`ziformula`) and dispersion (`dispformula`)
-formulae. For detailed examples of how to specify models, see the
+The formula argument for `cglmm()` is specified using the `{lme4}` style
+(for details see `vignette("lmer", package = "lme4")`). The only
+difference is that it allows for use of an `amp_acro()` call within the
+formula that is used to identify the circadian components and relevant
+variables in the `data.frame`. Any other combination of covariates can
+also be included in the formula as well as random effects and
+zero-inflation (`ziformula`) and dispersion (`dispformula`) formulae.
+For detailed examples of how to specify models, see the
 [mixed-models](https://rwparsons.github.io/GLMMcosinor/articles/mixed-models.html),
 [model-specification](https://rwparsons.github.io/GLMMcosinor/articles/model-specification.html)
 and
@@ -120,7 +120,9 @@ object
 
 The output shows the estimates for the raw coefficients in addition to
 the transformed estimates for amplitude (amp) and acrophase (acr) and
-MESOR (`(Intercept)`). The previous section of this vignette: [An overview of the statistical methods used for parameter estimation](https://rwparsons.github.io/GLMMcosinor/articles/GLMMcosinor.html#an-overview-of-the-statistical-methods-used-for-parameter-estimation)
+MESOR (`(Intercept)`). The previous section of this vignette: [An
+overview of the statistical methods used for parameter
+estimation](https://rwparsons.github.io/GLMMcosinor/articles/GLMMcosinor.html#an-overview-of-the-statistical-methods-used-for-parameter-estimation)
 outlines the difference between the raw coefficients and the transformed
 coefficients.
 
@@ -132,13 +134,13 @@ We would interpret the output as follows:
 
 -   `Acrophase estimate = 0.99913`
 
-Note that this estimate is in radians to align with conventions. To interpret 
-this, we can express `0.99913 radians` as a fraction of the total $2\pi$ and 
-multiply by the period to get the time when the response is a maximal. Hence,
-$\frac{0.99913}{2\pi} \times 12 = 1.908$ in the units of the `time_col` column
-in the original dataframe. This is saying that the peak response would occur
-after 1.908 time-units and every 12 time-units after this. We can confirm this
-by plotting:
+Note that this estimate is in radians to align with conventions. To
+interpret this, we can express `0.99913 radians` as a fraction of the
+total $2\pi$ and multiply by the period to get the time when the
+response is a maximal. Hence, $\frac{0.99913}{2\pi} \times 12 = 1.908$
+in the units of the `time_col` column in the original dataframe. This is
+saying that the peak response would occur after 1.908 time-units and
+every 12 time-units after this. We can confirm this by plotting:
 
 ```{r}
[…]
