### 2023-10-10 6fdd06c
created simulated dataset for DHARMa analysis
