+++ vignettes/model-specification.Rmd
@@ -410,28 +410,25 @@ appropriate. Firstly, we will compare the plots of both models using
 ```{r}
 library(GLMMcosinor)
 library(DHARMa)
-library(readr)
 
-mixed_mod_data <- as.data.frame(read_csv("data/mixed_mod_data.csv", 
-                                         show_col_types = FALSE))
-mixed_mod_data$mouse <-as.factor (mixed_mod_data$mouse) #make 'subject' a factor
+data(cosinor_mixed)
 
 #Assuming a Normal distribution 
-object_normal_mod <- GLMMcosinor::cglmm(Y ~ group + amp_acro(
-  time, 
+object_normal_mod <- GLMMcosinor::cglmm(Y ~ amp_acro(
+  times, 
   1, 
-  group = "group", 
   period = 24), 
-  data = mixed_mod_data)
+  data = cosinor_mixed)
 
 autoplot(object_normal_mod, predict.ribbon = FALSE, superimpose.data = TRUE)
 
 #Assuming a mixed Gamma distribution. 
-object_gamma_mixed_mod <- suppressWarnings(GLMMcosinor::cglmm(Y ~ group + amp_acro(time, 1, group = "group", period = 24)+ (1 + amp_acro1|mouse), data = mixed_mod_data, family = "Gamma"(link = 'log')))
+object_gamma_mixed_mod <- suppressWarnings(GLMMcosinor::cglmm(Y ~  amp_acro(times, 1, period = 24)+ (1 + amp_acro1|subject), data = cosinor_mixed, family = "Gamma"(link = 'log')))
+
 autoplot(object_gamma_mixed_mod, superimpose.data = TRUE)
 
 #If we wanted to generate regression fits for each of the subjects, we could use the following code:
-autoplot(object_gamma_mixed_mod, superimpose.data = TRUE, ranef_plot = 'mouse')
+autoplot(object_gamma_mixed_mod, superimpose.data = TRUE, ranef_plot = 'subject')
 ```
 
 Based on these figures, it seems that the Gamma model is a better fit.
