### 2023-10-10 e3a5036
run document() on cosinor_mixed data
