+++ R/data.R
@@ -11,3 +11,16 @@
 #' \code{X}.
 
 "vitamind"
+
+#' cosinor_mixed dataset for cosinor modelling examples.
+#'
+#' Simulated data set to illustrate a mixed cosinor model. The \code{Y} column
+#' contains a simulated outcome variable that varies over the time variable
+#' (\code{times}). The \code{subject} column is a grouping variable that can be
+#' used as a random effect. The rhythm has a period of 24 hours. Data was
+#' simulated using \code{simulate_cosinor}.
+#'
+#' @format A \code{data.frame} with 3 variables: \code{Y}, \code{times}, and
+#' \code{subject}.
+
+"cosinor_mixed"
