+++ man/cosinor_mixed.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{cosinor_mixed}
+\alias{cosinor_mixed}
+\title{cosinor_mixed dataset for cosinor modelling examples.}
+\format{
+A \code{data.frame} with 3 variables: \code{Y}, \code{times}, and
+\code{subject}.
+}
+\usage{
+cosinor_mixed
+}
+\description{
+Simulated data set to illustrate a mixed cosinor model. The \code{Y} column
+contains a simulated outcome variable that varies over the time variable
+(\code{times}). The \code{subject} column is a grouping variable that can be
+used as a random effect. The rhythm has a period of 24 hours. Data was
+simulated using \code{simulate_cosinor}.
+}
+\keyword{datasets}
