### 2023-10-17 cb431d7
(#2) make intro vignette and README clearer for beginners
