+++ vignettes/GLMMcosinor.Rmd
@@ -6,6 +6,7 @@ vignette: >
   %\VignetteIndexEntry{introduction-to-GLMMcosinor}
   %\VignetteEncoding{UTF-8}
   %\VignetteEngine{knitr::rmarkdown}
+bibliography: REFERENCES.bib
 editor_options: 
   markdown: 
     wrap: 72
@@ -18,89 +19,64 @@ knitr::opts_chunk$set(
 )
 ```
 
-```{r setup}
-library(GLMMcosinor)
-library(dplyr)
-withr::with_seed(
-  50,
-  {
-    testdata_simple <- simulate_cosinor(1000,
-      n_period = 2,
-      mesor = 5,
-      amp = 2,
-      acro = 1,
-      beta.mesor = 4,
-      beta.amp = 1,
-      beta.acro = 0.5,
-      family = "poisson",
-      period = c(12),
-      n_components = 1,
-      beta.group = TRUE
-    )
-
-    testdata_two_components <- simulate_cosinor(1000,
-      n_period = 10,
-      mesor = 7,
-      amp = c(0.1, 0.4),
-      acro = c(1, 1.5),
-      beta.mesor = 4.4,
-      beta.amp = c(2, 1),
-      beta.acro = c(1, -1.5),
-      family = "poisson",
-      period = c(12, 6),
-      n_components = 2,
-      beta.group = TRUE
-    )
-
-    testdata_multicomponent <- simulate_cosinor(1000,
-      n_period = 10,
-      mesor = 7,
-      amp = c(0.1, 0.4, 0.5),
-      acro = c(1, 1.5, 0.1),
-      beta.mesor = 4.4,
-      beta.amp = c(2, 1, 0.4),
-      beta.acro = c(1, -1.5, -1),
-      family = "poisson",
-      period = c(12, 6, 12),
-      n_components = 3,
-      beta.group = TRUE
-    )
-  }
-)
-```
-
 ```{r srr, eval = FALSE, echo = FALSE}
 #' @srrstats {G1.0}
 #' @srrstats {G1.1}
 #' @srrstats {G1.3}
 ```
 
-```{r, message=F, warning=F}
-data(vitamind)
-object <- cglmm(vit_d ~ X + amp_acro(time,
-  n_components = 1, period = 12,
-  group = "X"
-), data = vitamind)
+## An brief introduction to the cosinor model
 
-autoplot(object,
-  superimpose.data = TRUE,
-  x_str = "X",
-  predict.ribbon = FALSE,
-  data_opacity = 0.2
-)
-```
+A cosinor model aims to model the amplitude ($A$), acrophase ($\phi$), 
+and MESOR ($M$) of a rhythmic dataset.
+
+-   MESOR ($M$) is the Midline Estimating Statistic of Rhythm, and may
+    also be referred to as the equilibrium point.
+
+-   Amplitude ($A$) is the difference between the MESOR and the maximum height 
+    of the rhythm.
+
+-   Acrophase ($\phi$) is the phase at which the maximal response
+    occurs.
+
+These could be modelled using a cosine function:
+
+$$Y(t) = M + Acos(\frac{2\pi t}{\tau} - \phi) + e(t)$$ where $e(t)$ is
+the error term.
+
+However, these cannot be estimated using a linear modelling framework! Other
+packages, including 
+[`{circacompare}`](https://cran.r-project.org/package=circacompare) 
+[@parsons2020circacompare], fit this exact
+nonlinear model but most packages (including this one) decomposes this into 
+linear terms, creating the cosinor model:
+
+$$Y(t) = M + \beta x + \gamma z + e(t)$$
+
[…]
