+++ vignettes/REFERENCES.bib
@@ -1,0 +1,21 @@
+@article{cornelissen2014cosinor,
+  title={Cosinor-based rhythmometry},
+  author={Cornelissen, Germaine},
+  journal={Theoretical Biology and Medical Modelling},
+  volume={11},
+  number={1},
+  pages={1--24},
+  year={2014},
+  publisher={BioMed Central}
+}
+
+@article{parsons2020circacompare,
+  title={CircaCompare: a method to estimate and statistically support differences in mesor, amplitude and phase, between circadian rhythms},
+  author={Parsons, Rex and Parsons, Richard and Garner, Nicholas and Oster, Henrik and Rawashdeh, Oliver},
+  journal={Bioinformatics},
+  volume={36},
+  number={4},
+  pages={1208--1212},
+  year={2020},
+  publisher={Oxford University Press}
+}
