### 2023-10-17 6cddb16
add tests for plotting single component, non-grouped model
