+++ vignettes/multiple-components.Rmd
@@ -15,7 +15,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-```{r setup}
+```{r, echo=F}
 library(GLMMcosinor)
 withr::with_seed(
   50,
@@ -48,7 +48,6 @@ withr::with_seed(
       beta.group = TRUE
     )
 
-
     testdata_three_components <- simulate_cosinor(1000,
       n_period = 2,
       mesor = 7,
@@ -75,6 +74,25 @@ To generate a multi-component model, set `n_components` in the `amp_acro()` part
 The `period` argument must also match the length of `n_components`, where the order of the periods corresponds to their assigned component. For example, if `n_components = 2`, and `period = c(12,6)`, then the first component has a `period` of 12 and the second a period of 6. Similarly to the `group` argument, if only one period is supplied despite `n_components` being greater than 1, then this period will be matched to each component.
 
 For example:
+
+```{r, eval=F}
+library(GLMMcosinor)
+testdata_two_components <- simulate_cosinor(
+  1000,
+  n_period = 10,
+  mesor = 7,
+  amp = c(0.1, 0.4),
+  acro = c(1, 1.5),
+  beta.mesor = 4.4,
+  beta.amp = c(2, 1),
+  beta.acro = c(1, -1.5),
+  family = "poisson",
+  period = c(12, 6),
+  n_components = 2,
+  beta.group = TRUE
+)
+```
+
 
 ```{r, message=F, warning=F}
 object <- cglmm(
@@ -109,6 +127,23 @@ autoplot(object)
 If a multicomponent model has one component that is grouped with other components that aren't, the vector input for `group` must still be the same length as `n_components` but have the non-grouped components represented as `group = NA`.
 
 For example, if wanted only the first component to have a grouped component, we would specify the `group` argument as `group = c("group", NA))` . Here, the first component is grouped by `group`, and the second component is not grouped. The data was simulated such that the second component was the same for both groups.
+
+```{r, eval=F}
+testdata_two_components_grouped <- simulate_cosinor(
+  1000,
+  n_period = 5,
+  mesor = 3.7,
+  amp = c(0.1, 0.4),
+  acro = c(1, 1.5),
+  beta.mesor = 4,
+  beta.amp = c(2, 0.4),
+  beta.acro = c(1, 1.5),
+  family = "poisson",
+  period = c(12, 6),
+  n_components = 2,
+  beta.group = TRUE
+)
+```
 
 ```{r, message=F, warning=F}
 object <- cglmm(
@@ -193,6 +228,24 @@ cglmm(
 ## Generating a three-component model
 
 The plot below shows a 3-component model with the simulated data overlayed:
+
+```{r, eval=F}
+testdata_three_components <- simulate_cosinor(
+  1000,
+  n_period = 2,
+  mesor = 7,
+  amp = c(0.1, 0.4, 0.5),
+  acro = c(1, 1.5, 0.1),
+  beta.mesor = 4.4,
+  beta.amp = c(2, 1, 0.4),
+  beta.acro = c(1, -1.5, -1),
+  family = "poisson",
+  period = c(12, 6, 12),
+  n_components = 3,
+  beta.group = TRUE
+)
+```
+
 
 ```{r, message=F, warning=F}
 object <- cglmm(
