+++ vignettes/model-specification.Rmd
@@ -27,11 +27,14 @@ knitr::opts_chunk$set(
 ```{r setup}
 library(GLMMcosinor)
 library(dplyr)
+```
 
+```{r, echo=F}
 withr::with_seed(
   50,
   {
-    testdata_simple <- simulate_cosinor(1000,
+    testdata_simple <- simulate_cosinor(
+      1000,
       n_period = 2,
       mesor = 5,
       amp = 2,
@@ -45,7 +48,8 @@ withr::with_seed(
       beta.group = TRUE
     )
 
-    testdata_simple_gaussian <- simulate_cosinor(1000,
+    testdata_simple_gaussian <- simulate_cosinor(
+      1000,
       n_period = 2,
       mesor = 5,
       amp = 2,
@@ -59,7 +63,8 @@ withr::with_seed(
       beta.group = TRUE
     )
 
-    testdata_two_components <- simulate_cosinor(1000,
+    testdata_two_components <- simulate_cosinor(
+      1000,
       n_period = 10,
       mesor = 7,
       amp = c(0.1, 0.4),
@@ -75,6 +80,7 @@ withr::with_seed(
   }
 )
 ```
+
 
 ## `cglmm()`
 
@@ -107,7 +113,26 @@ Here, we fit a simple cosinor model to "testdata_simple" - simulated
 data from a Poisson distribution loaded in this vignette. In this
 example, there is no grouping variable.
 
+```{r, eval = F}
+testdata_simple <- simulate_cosinor(
+  1000,
+  n_period = 2,
+  mesor = 5,
+  amp = 2,
+  acro = 1,
+  beta.mesor = 4,
+  beta.amp = 1,
+  beta.acro = 0.5,
+  family = "poisson",
+  period = c(12),
+  n_components = 1,
+  beta.group = TRUE
+)
+```
+
 ```{r, message=F, warning=F}
+
+
 object <- cglmm(
   Y ~ amp_acro(times,
     period = 12
@@ -151,7 +176,26 @@ autoplot(object, superimpose.data = TRUE)
 Now, we can add a grouping variable by adding the name of the group in
 the `amp_acro()` function:
 
+```{r, eval=F}
+testdata_simple_gaussian <- simulate_cosinor(
+  1000,
+  n_period = 2,
+  mesor = 5,
+  amp = 2,
+  acro = 1,
+  beta.mesor = 4,
+  beta.amp = 1,
+  beta.acro = 0.5,
+  family = "gaussian",
+  period = c(12),
+  n_components = 1,
+  beta.group = TRUE
+)
+```
+
 ```{r, message=F, warning=F}
+
+
 object <- cglmm(
   Y ~ amp_acro(times,
     period = 12,
@@ -261,6 +305,24 @@ value of `n_components` will need to match the length of the `group` and
 `period` arguments as these will be combined for each component.
 
 For example:
+
+```{r, eval=F}
+testdata_two_components <- simulate_cosinor(
+  1000,
+  n_period = 10,
+  mesor = 7,
+  amp = c(0.1, 0.4),
+  acro = c(1, 1.5),
+  beta.mesor = 4.4,
+  beta.amp = c(2, 1),
+  beta.acro = c(1, -1.5),
+  family = "poisson",
+  period = c(12, 6),
+  n_components = 2,
+  beta.group = TRUE
+)
+```
+
 
 ```{r, message=F, warning=F}
 cglmm(
@@ -395,52 +457,14 @@ zero-inflation models.*
 
 ## Assessing residual diagnostics of `cglmm` regression models using DHARMa
 
-DHARMa is an R package used to assess residual diagnostics of regression
-models, including the generalised linear models generated by `cglmm()`
-[REFERENCE]. This will allow us to evaluate the appropriateness of
-specifying different models.
[…]

[…]
