### 2023-11-01 e6d68fc
add a failing (commented out) test: autoplot with model including covariates
