+++ DESCRIPTION
@@ -7,7 +7,11 @@ Authors@R: c(
            comment = c(ORCID = "0000-0002-6053-8174")),
     person("Oliver", "Jayasinghe", role = "aut"),
     person("Nicole", "White", role = "aut",
-           comment = c(ORCID = "0000-0002-9292-0773"))
+           comment = c(ORCID = "0000-0002-9292-0773")),
+    person("Oliver", "Rawashdeh", role = c("aut", "fnd"),
+           comment = c(ORCID = "0000-0002-7147-4778")),
+    person("Prasad", "Chunduri", role = c("ctb", "fnd"),
+           comment = c(ORCID = "0000-0001-7297-7580"))
   )
 Description: Fit the cosinor model using the glmmTMB framework, allowing
     for a wide range of link functions and capacity of hierarchical
