### 2023-11-02 1e7c739
add 'cov_list' arg to autoplot() and implement associated tests
