### 2023-11-02 39bb81f
add 'cov_list' as an argument in autoplot() and implement associated tests
