+++ R/predict.R
@@ -31,12 +31,6 @@ predict.cglmm <- function(object, newdata, ...) {
     return(stats::predict(object$fit, ...))
   }
 
-  # all(names(newdata) %in% names(object$fit$frame))
-  # TODO: modify dataset to include NA in random effects columns for which you
-  # want a population estimate instead of a random effect. Population effect by
-  # default, and the user should specify which variables they want as random
-  # effects. This should be done in the autoplot script
-
   # pass new dataset that's being used for prediction in this function
   nd <- update_formula_and_data(
     data = newdata,
