### 2023-11-02 05de966
make col_list work if user specifies some but not all covariates
