### 2023-11-02 6e18673
Update autoplot.cglmm.Rd
