### 2023-11-06 f08fc5f
Fix bug in ranef plots that would happen when random effect variables are not factors in df
