### 2023-11-17 8f75e2a
update test_cosinor_components so that it works when there is no group arg specified
This change increases flexibility within the GLMMCosinor app
