### 2023-11-20 f2e9480
update labelling in polar_plots
