+++ R/polar_plot.R
@@ -63,6 +63,8 @@
 #' polar plots.
 #' @param quietly Analogous to verbose, this \code{logical} argument controls
 #' whether messages are displayed in the console.
+#' @param show_component_labels Logical argument, TRUE by default. When TRUE,
+#' the polar plots have labels corresponding to their components.
 #' @param ... Additional, ignored arguments.
 #'
 #' @srrstats {G1.4}
@@ -81,7 +83,7 @@
 polar_plot <- function(x,
                        ci_level = 0.95,
                        n_breaks = 5,
-                       component_index,
+                       component_index = NULL,
                        grid_angle_segments = 8,
                        radial_units = c("radians", "degrees", "period"),
                        clockwise = FALSE,
@@ -94,6 +96,7 @@ polar_plot <- function(x,
                        view = c("full", "zoom", "zoom_origin"),
                        overlay_parameter_info = FALSE,
                        quietly = TRUE,
+                       show_component_labels = TRUE,
                        ...) {
   UseMethod("polar_plot")
 }
@@ -179,7 +182,7 @@ polar_plot <- function(x,
 polar_plot.cglmm <- function(x,
                              ci_level = 0.95,
                              n_breaks = 5,
-                             component_index,
+                             component_index = NULL,
                              grid_angle_segments = 8,
                              radial_units = c(
                                "radians",
@@ -205,10 +208,11 @@ polar_plot.cglmm <- function(x,
                              ),
                              overlay_parameter_info = FALSE,
                              quietly = TRUE,
+                             show_component_labels = TRUE,
                              ...) {
   # checking the quality of inputs
   assertthat::assert_that(inherits(x, "cglmm"),
-    msg = "'x' must be of class cglmm"
+                          msg = "'x' must be of class cglmm"
   )
 
   validate_ci_level(ci_level)
@@ -222,13 +226,19 @@ polar_plot.cglmm <- function(x,
     msg = "'grid_angle_segments' must be an integer greater than 0"
   )
   assertthat::assert_that(is.logical(quietly),
-    msg = "'quietly' must a logical argument, either TRUE or FALSE"
+                          msg = "'quietly' must a logical argument, either TRUE or FALSE"
   )
   assertthat::assert_that(is.logical(clockwise),
-    msg = "'clockwise' must be a logical argument, either TRUE or FALSE "
+                          msg = "'clockwise' must be a logical argument, either TRUE or FALSE "
   )
+
+  assertthat::assert_that(is.logical(show_component_labels),
+                          msg = paste("'show_component_labels' must be a",
+                                      " logical argument, either TRUE or FALSE")
+  )
+
   assertthat::assert_that(is.numeric(text_size) & text_size > 0,
-    msg = "'text_size' must be a number greater than 0"
+                          msg = "'text_size' must be a number greater than 0"
   )
   assertthat::assert_that(
     is.numeric(text_opacity) & text_opacity >= 0 & text_opacity <= 1,
@@ -252,16 +262,16 @@ polar_plot.cglmm <- function(x,
     )
   }
   assertthat::assert_that(is.character(circle_linetype),
-    msg = paste(
-      "'circle_linetype' must be a character. See ?linetype",
-      "for more details"
-    )
+                          msg = paste(
[…]
