### 2023-11-20 d3c4f84
update documentation to pass check()
