### 2023-11-20 0781fef
Capitalize 'Component' in polar_plot() output
