### 2023-11-22 b93617a
update vignettes and allow strings for time_col input
