+++ vignettes/GLMMcosinor.Rmd
@@ -27,14 +27,14 @@ knitr::opts_chunk$set(
 
 ## An brief introduction to the cosinor model
 
-A cosinor model aims to model the amplitude ($A$), acrophase ($\phi$), 
+A cosinor model aims to model the amplitude ($A$), acrophase ($\phi$),
 and MESOR ($M$) of a rhythmic dataset.
 
 -   MESOR ($M$) is the Midline Estimating Statistic of Rhythm, and may
     also be referred to as the equilibrium point.
 
--   Amplitude ($A$) is the difference between the MESOR and the maximum height 
-    of the rhythm.
+-   Amplitude ($A$) is the difference between the MESOR and the maximum
+    height of the rhythm.
 
 -   Acrophase ($\phi$) is the phase at which the maximal response
     occurs.
@@ -44,39 +44,41 @@ These could be modelled using a cosine function:
 $$Y(t) = M + Acos(\frac{2\pi t}{\tau} - \phi) + e(t)$$ where $e(t)$ is
 the error term.
 
-However, these cannot be estimated using a linear modelling framework! Other
-packages, including 
-[`{circacompare}`](https://cran.r-project.org/package=circacompare) 
-[@parsons2020circacompare], fit this exact
-nonlinear model but most packages (including this one) decomposes this into 
-linear terms, creating the cosinor model:
+However, these cannot be estimated using a linear modelling framework!
+Other packages, including
+[`{circacompare}`](https://cran.r-project.org/package=circacompare)
+[@parsons2020circacompare], fit this exact nonlinear model but most
+packages (including this one) decomposes this into linear terms,
+creating the cosinor model:
 
 $$Y(t) = M + \beta x + \gamma z + e(t)$$
 
 Where $x =cos(\frac{2\pi t}{τ})$, $z =sin(\frac{2\pi t}{τ})$,
 $\beta = A cos(\phi)$, $\gamma = A sin(\phi)$
 
-These original parameters can be recovered: for amplitude ($A$) and acrophase 
-($\phi$), the estimates for $\hat\beta$ and $\hat\gamma$ must be transformed 
-as per the following equations:
+These original parameters can be recovered: for amplitude ($A$) and
+acrophase ($\phi$), the estimates for $\hat\beta$ and $\hat\gamma$ must
+be transformed as per the following equations:
 
 $$\hat\phi = \arctan(\frac{\hat\gamma}{\hat\beta}) $$
 
-$$\hat A = \sqrt{\hat\beta ^2 + \hat\gamma ^ 2}$$
-For a more thorough introduction to cosinor modelling, see 
-[here](https://doi.org/10.1186%2F1742-4682-11-16) [@cornelissen2014cosinor].
+$$\hat A = \sqrt{\hat\beta ^2 + \hat\gamma ^ 2}$$ For a more thorough
+introduction to cosinor modelling, see
+[here](https://doi.org/10.1186%2F1742-4682-11-16)
+[@cornelissen2014cosinor].
 
 ## Introduction
 
-`{GLMMcosinor}` allows the user to fit generalised linear models based on 
-rhythmic data with a cosinor model. It allows users to  summarise, predict, and
-plot these models too. Existing packages have focused primarily on Gaussian
-data. Some circadian regression modelling packages have allowed users to specify
-generalised linear models, but with limited flexibility. `{GLMMcosinor}` takes
-a comprehensive approach to modelling by harnessing the `{glmmTMB}` package, 
-that has a wide range of available link functions, allowing users to model 
-rhythmic data from a wide range of distributions (for full list - see `?family`
-and `?glmmTMB::family_glmmTMB`) including:
+`{GLMMcosinor}` allows the user to fit generalised linear models based
+on rhythmic data with a cosinor model. It allows users to summarise,
[…]
