### 2023-11-24 147cc72
fix bug when specifying radial_units = "period" and overlay_parameter = TRUE
