### 2023-11-24 f218b57
update documentation for circle_linetype argument in polar_plot.R
