### 2023-12-04 e576b82
provide brief explanation of how glmmTMB works, with references
