+++ vignettes/model-specification.Rmd
@@ -81,7 +81,6 @@ withr::with_seed(
 )
 ```
 
-
 ## `cglmm()`
 
 `cglmm()` wrangles the data appropriately to fit the cosinor model given
@@ -323,7 +322,6 @@ testdata_two_components <- simulate_cosinor(
 )
 ```
 
-
 ```{r, message=F, warning=F}
 cglmm(
   Y ~ group + amp_acro(
@@ -457,11 +455,12 @@ zero-inflation models.*
 
 ## Assessing residual diagnostics of `cglmm` regression models using DHARMa
 
-`{DHARMa}` is an R package used to assess residual diagnostics of regression
-models fit using `{glmmTMB}` (which is what is used by `cglmm()`).
+`{DHARMa}` is an R package used to assess residual diagnostics of
+regression models fit using `{glmmTMB}` (which is what is used by
+`cglmm()`).
 
-For example, we can apply the functions from `DHARMa` on the `glmmTMB` model
-within by accessing it with `$fit`.
+For example, we can apply the functions from `DHARMa` on the `glmmTMB`
+model within by accessing it with `$fit`.
 
 ```{r}
 library(DHARMa)
