### 2023-12-06 4076ae0
split polar_plot code into functions
one function generates the background, and the other generates the estimates and ellipses. This means that it is easier to iteratively add plot layers within the code. Also, added arguments: xlims and ylims to polar_plot to allow greater customisation
