+++ R/polar_plot.R
@@ -149,6 +149,8 @@ polar_plot <- function(x,
 #' the \code{linetype} of the radial circles in background of the polar plot.
 #' See \code{?linetype}
 #' for more details.
+#' @param xlims A vector of length two containing the limits for the x-axis.
+#' @param ylims A vector of length two containing the limits for the y-axis.
 #' @param start A \code{character}, within
 #' \code{c("right", "left", "top", "bottom")} that determines where angle 0 is
 #' located. If \code{start = "top"}, and \code{clockwise = TRUE}, the angle
@@ -213,6 +215,8 @@ polar_plot.cglmm <- function(x,
                              overlay_parameter_info = FALSE,
                              quietly = TRUE,
                              show_component_labels = TRUE,
+                             xlims,
+                             ylims,
                              ...) {
   # checking the quality of inputs
   assertthat::assert_that(inherits(x, "cglmm"),
@@ -265,6 +269,35 @@ polar_plot.cglmm <- function(x,
       )
     )
   }
+
+  if (!missing(xlims)) {
+    assertthat::assert_that(
+      length(xlims) == 2 & is.numeric(xlims) & xlims[1] < xlims[2],
+      msg = paste(
+        "'xlims' must be a vector with the first element being the",
+        "lower x coordinate, and the second being the upper",
+        "x coordinate"
+      )
+    )
+    xlims_check <- TRUE
+  } else {
+    xlims_check <- FALSE
+  }
+
+  if (!missing(ylims)) {
+    assertthat::assert_that(
+      length(xlims) == 2 & is.numeric(ylims) & ylims[1] < ylims[2],
+      msg = paste(
+        "'ylims' must be a vector with the first element being the",
+        "lower y coordinate, and the second being the upper",
+        "y coordinate"
+      )
+    )
+    ylims_check <- TRUE
+  } else {
+    ylims_check <- FALSE
+  }
+
   assertthat::assert_that(is.character(circle_linetype) || is.numeric(circle_linetype),
     msg = paste(
       "'circle_linetype' must be a character or numeric. See ?linetype",
@@ -501,18 +534,96 @@ polar_plot.cglmm <- function(x,
       group_level_colour_index <- length(group_level)
     }
 
-    plot_obj <-
-      ggplot2::ggplot() +
-      ggforce::geom_circle(
-        ggplot2::aes(
-          x0 = 0,
-          y0 = 0,
-          r = scales::breaks_pretty(n = n_breaks)(c(0, max_plot_radius))
-        ),
-        alpha = 0.01,
-        linetype = circle_linetype
-      ) +
-      ggforce::geom_ellipse(
+    # generates the background grid for the polar plot
+    get_background_grid <- function(n_breaks,
+                                    max_plot_radius,
+                                    circle_linetype,
+                                    dial_pos_full_x,
+                                    dial_pos_full_y,
+                                    time_labels,
+                                    text_size,
+                                    text_opacity,
+                                    contour_labels,
+                                    grid_angle_segments
+                                    ){
+      plot_background <- ggplot2::ggplot() +
+        ggforce::geom_circle( #plots the background circles
+          ggplot2::aes(
+            x0 = 0,
+            y0 = 0,
+            r = scales::breaks_pretty(n = n_breaks)(c(0, max_plot_radius))
+          ),
+          alpha = 0.01,
+          linetype = circle_linetype
+        ) +
+        ggplot2::geom_segment( #plots the background grid
+          ggplot2::aes(
+            x = dial_pos_full_x,
+            y = dial_pos_full_y,
+            xend = -dial_pos_full_x,
+            yend = -dial_pos_full_y
+          ),
+          linetype = 10,
+          alpha = 0.4
+        ) +
[…]
