### 2023-12-06 f04a94c
add arguments xlims and ylims to polar_plot parent function
