### 2023-12-06 a958853
update polar_plot documentation
