### 2023-12-07 da57422
fix bug in autoplot() when time is specified as an interaction term
