### 2023-12-10 ee7a386
Update README to inlcude a brief explanation of 'glmmTMB' framework
