### 2023-12-11 7467f9b
avoid all uses of native pipe (|>)
