### 2023-12-11 91cbaed
fix up references in README
