+++ vignettes/REFERENCES.bib
@@ -19,3 +19,23 @@
   year={2020},
   publisher={Oxford University Press}
 }
+
+@Manual{sachs2023cosinor,
+  title = {cosinor: Tools for Estimating and Predicting the Cosinor Model},
+  author = {Michael Sachs},
+  year = {2023},
+  note = {R package version 1.2.3},
+  url = {https://CRAN.R-project.org/package=cosinor},
+}
+
+@article{Kristensen2016TMB,
+ title={TMB: Automatic Differentiation and Laplace Approximation},
+ volume={70},
+ url={https://www.jstatsoft.org/index.php/jss/article/view/v070i05},
+ doi={10.18637/jss.v070.i05},
+ number={5},
+ journal={Journal of Statistical Software},
+ author={Kristensen, Kasper and Nielsen, Anders and Berg, Casper W. and Skaug, Hans and Bell, Bradley M.},
+ year={2016},
+ pages={1–21}
+}
