### 2023-12-14 bec04d3
create sigma() method to get dispersion parameter
