+++ man/reexports.Rd
@@ -1,9 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/autoplot.R
+% Please edit documentation in R/autoplot.R, R/sigma.R
 \docType{import}
 \name{reexports}
 \alias{reexports}
 \alias{autoplot}
+\alias{sigma}
 \title{Objects exported from other packages}
 \keyword{internal}
 \description{
@@ -12,5 +13,7 @@ below to see their documentation.
 
 \describe{
   \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
+
+  \item{stats}{\code{\link[stats]{sigma}}}
 }}
 
