### 2023-12-14 RWParsons/GLMMcosinor#2 @RWParsons
Suggestions and how they were addressed

- [x] The package docs are quite impressive to someone who is familiar with the cosinor model, but maybe it is not so accessible to someone who is not. To remedy this, I think the vignettes and possibly the readme could use a more thorough description of the motivation for using the cosinor model. Maybe in the context of the vitamind example dataset you could give a short explanation of the biological rationale for the model, what the potential research questions that could be answered, and how the data were (hypothetically) collected.
> Have updated the vignettes and README to make this more clear, particularly the getting-started vignette.
- [x]  ~~Feature suggestion: predict function could be enhanced to also obtain predictions of time-independent parameters given a vector of covariates, e.g., the amplitude, acrophase, or MESOR.~~
> Discussed in [stats-review issue](https://github.com/ropensci/software-review/issues/603#issuecomment-1747303693). Will keep this in mind for future versions of the package but is not currently in scope for this initial version. A future version that can handle time-varying effects on amplitude and phase.
- [x]  The title in DESCRIPTION will need to be in title case if you plan to submit to CRAN.
> Have updated title.
- [x] Inconsistent use of snake case and dot case in the function arguments of: autoplot, simulate_cosinor
> Have updated so that all function names are snake_case and S3 class names as camelCase.
- [x] The vitamind example dataset description is completely generic (e.g., “Y is an outcome variable”). You could consider adding some fake details about the measurements, or instead changing the name of the dataset to something equally generic, like example_data or similar.
> Have updated dataset so that the response variable is `vit_d`, not `Y`.
