### 2023-12-14 #603 @RWParsons
Thanks @Paula-Moraga and thank you @sachsmc and @jcavieresg for your helpful and constructive reviews of the package.

I made an issues on the GLMMcosinor repo with the suggestions from each reviewer.

@jcavieresg's review issue: https://github.com/RWParsons/GLMMcosinor/issues/4

@sachsmc's review issue: https://github.com/RWParsons/GLMMcosinor/issues/2

In each of these issues, I have left a comment on how we addressed each suggestion (and checked them off as they were completed).
