### 2023-12-18 7fdb406
update references in vignette
