+++ vignettes/REFERENCES.bib
@@ -39,3 +39,17 @@
  year={2016},
  pages={1–21}
 }
+
+@article{brooks2017glmmTMB,
+  author = {Mollie E. Brooks and Kasper Kristensen and Koen J. {van
+    Benthem} and Arni Magnusson and Casper W. Berg and Anders Nielsen
+    and Hans J. Skaug and Martin Maechler and Benjamin M. Bolker},
+  title = {{glmmTMB} Balances Speed and Flexibility Among Packages
+    for Zero-inflated Generalized Linear Mixed Modeling},
+  year = {2017},
+  journal = {The R Journal},
+  doi = {10.32614/RJ-2017-066},
+  pages = {378--400},
+  volume = {9},
+  number = {2},
+}
