+++ vignettes/fixtures/methods-table.R
@@ -145,9 +145,11 @@ dat <- flextable::color(
   col = function(x) ifelse(x == "&#x2713;", "green", "red")
 )
 dat <- ftExtra::colformat_md(dat)
-dat <- flextable::theme_zebra(dat,
-                              even_header = "#FFFFFF",
-                              even_body = "#FFFFFF")
+dat <- flextable::theme_zebra(
+  dat,
+  even_header = "#FFFFFF",
+  even_body = "#FFFFFF"
+)
 dat <- flextable::set_header_labels(
   dat,
   values = list(
@@ -163,4 +165,3 @@ dat <- flextable::set_header_labels(
     ref = "Reference"
   )
 )
-
