+++ vignettes/mixed-models.Rmd
@@ -28,9 +28,6 @@ library(ggplot2)
 To illustrate an example of using a model with random effects on the cosinor components, we will first simulate some data with `id`-level differences in amplitude and acrophase.
 
 ```{r}
-  withr::with_seed(
-    50,
-    {
 f_sample_id <- function(id_num,
                         n = 30,
                         mesor,
@@ -67,12 +64,32 @@ dat_mixed <- do.call(
       n_components = 1
     )
   })
-) 
-  dat_mixed$subject <- as.factor(dat_mixed$subject)
-  
-    })
-
+)
+dat_mixed$subject <- as.factor(dat_mixed$subject)
 ```
+
+```{r echo=FALSE}
+withr::with_seed(
+  50,
+  {
+    dat_mixed <- do.call(
+      "rbind",
+      lapply(1:30, function(x) {
+        f_sample_id(
+          id_num = x,
+          mesor = rnorm(1, mean = 0, sd = 1),
+          amp = rnorm(1, mean = 3, sd = 0.5),
+          acro = rnorm(1, mean = 1.5, sd = 0.2),
+          period = 24,
+          n_components = 1
+        )
+      })
+    )
+    dat_mixed$subject <- as.factor(dat_mixed$subject)
+  }
+)
+```
+
 
 A quick graph shows how there are individual differences in terms of MESOR, amplitude and phase.
 
@@ -89,14 +106,22 @@ For the model, we should include a random effect for the MESOR, amplitude and ac
 
 In the model formula, we can use the special `amp_acro[n]` which represents the n<sup>th</sup> cosinor component. In this case, we only have one component so we use `amp_acro1`. Following the `{lme4}`-style mixed model formula, we add our random effect for this component and the intercept term (MESOR) clustered within subjects by using `(1 + amp_acro1 | subject)`. The code below fits this model
 
-```{r}
+```r
 mixed_mod <- cglmm(
-  Y ~ amp_acro(times,
-    n_components = 1,
-    period = 24
-  ) + (1 + amp_acro1 | subject),
-  data = dat_mixed
-)
+    Y ~ amp_acro(times, n_components = 1, period = 24) + 
+      (1 + amp_acro1 | subject),
+    data = dat_mixed
+  )
+```
+
+```{r echo=FALSE}
+withr::with_seed(42, {
+  mixed_mod <- cglmm(
+    Y ~ amp_acro(times, n_components = 1, period = 24) +
+      (1 + amp_acro1 | subject),
+    data = dat_mixed
+  )
+})
 ```
 
 This works by replacing the amp_acro1 with the relevant cosinor components when the data is rearranged and the formula created. The formula created can be accessed using `.$formula`, and shows the `amp_acro1` is replaced by the `main_rrr1` and `main_sss1` (the cosine and sine components of time that also appear in the fixed effects).
