+++ vignettes/model-visualisations.Rmd
@@ -34,7 +34,7 @@ withr::with_seed(42, {
     period = c(12, 6),
     n_components = 2
   )
-  
+
   testdata_period_diff <- simulate_cosinor(
     1000,
     n_period = 1,
@@ -181,15 +181,15 @@ argument which is 20 by default.
 
 ```{r, eval=F}
 testdata_period_diff <- simulate_cosinor(
-    1000,
-    n_period = 1,
-    mesor = 7,
-    amp = c(0.1, 0.4),
-    acro = c(1, 1.5),
-    family = "poisson",
-    period = c(12, 1000),
-    n_components = 2
-  )
+  1000,
+  n_period = 1,
+  mesor = 7,
+  amp = c(0.1, 0.4),
+  acro = c(1, 1.5),
+  family = "poisson",
+  period = c(12, 1000),
+  n_components = 2
+)
 ```
 
 ```{r, message=F, warning=F}
