### 2023-12-27 616e1cb
update amp_acro to accept string argument for time_col
This change was made several months ago
