### 2023-12-27 16d7cee
Revert "update amp_acro to accept string argument for time_col"
This reverts commit 616e1cb9717783f1d9246bbd28463c091319b7a8.
