+++ tests/testthat/_snaps/summary.cosinor.md
@@ -16,9 +16,9 @@
                 estimate standard.error lower.CI upper.CI p.value    
     (Intercept)    30.33           0.38    29.58     31.1  <2e-16 ***
     [X=0]:amp1      6.30           0.68     4.97      7.6  <2e-16 ***
-    [X=1]:amp1      7.98           1.12     5.79     10.2   9e-13 ***
+    [X=1]:amp1      7.98           0.91     6.20      9.8  <2e-16 ***
     [X=0]:acr1      1.43           0.10     1.24      1.6  <2e-16 ***
-    [X=1]:acr1      0.62           0.14     0.35      0.9   8e-06 ***
+    [X=1]:acr1      0.62           0.12     0.39      0.9   1e-07 ***
     ---
     Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
 
@@ -40,9 +40,9 @@
                 estimate standard.error lower.CI upper.CI p.value    
     (Intercept)   30.328          0.375   29.593     31.1  <2e-16 ***
     [X=0]:amp1     6.351          0.697    4.985      7.7  <2e-16 ***
-    [X=1]:amp1     8.030          1.130    5.814     10.2   1e-12 ***
-    [X=0]:acr1     1.420          0.098    1.228      1.6  <2e-16 ***
-    [X=1]:acr1     0.607          0.141    0.330      0.9   2e-05 ***
+    [X=1]:amp1     8.030          0.926    6.214      9.8  <2e-16 ***
+    [X=0]:acr1     1.420          0.098    1.227      1.6  <2e-16 ***
+    [X=1]:acr1     0.607          0.123    0.365      0.8   8e-07 ***
     ---
     Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
     
@@ -63,9 +63,9 @@
                 estimate standard.error lower.CI upper.CI p.value    
     (Intercept)    3.313          0.100    3.116      3.5  <2e-16 ***
     [X=0]:amp1     0.392          0.173    0.053      0.7    0.02 *  
-    [X=1]:amp1     0.100          0.315   -0.518      0.7    0.75    
-    [X=0]:acr1     0.269          0.480   -0.673      1.2    0.58    
-    [X=1]:acr1     2.539          3.174   -3.681      8.8    0.42    
+    [X=1]:amp1     0.100          0.243   -0.376      0.6    0.68    
+    [X=0]:acr1     0.269          0.521   -0.753      1.3    0.61    
+    [X=1]:acr1     2.539          2.379   -2.125      7.2    0.29    
     ---
     Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
     
@@ -84,7 +84,7 @@
                 estimate standard.error lower.CI upper.CI p.value
     (Intercept) -2.4e+01        1.3e+04 -2.6e+04    26409       1
     [X=0]:amp1   3.7e+00        1.6e+04 -3.1e+04    30621       1
-    [X=1]:amp1   4.7e-01        2.8e+04 -5.4e+04    54265       1
-    [X=0]:acr1   2.7e+00        3.6e+03 -7.1e+03     7144       1
-    [X=1]:acr1   1.2e+00        6.1e+04 -1.2e+05   118857       1
+    [X=1]:amp1   4.7e-01        2.6e+04 -5.0e+04    50158       1
+    [X=0]:acr1   2.7e+00        2.7e+03 -5.2e+03     5228       1
+    [X=1]:acr1   1.2e+00        5.8e+04 -1.1e+05   114240       1
 
