### 2023-12-28 12f9eac
update documentation to pass check()
