+++ man/polar_plot.Rd
@@ -1,9 +1,32 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/polar_plot.R
+% Please edit documentation in R/polar_plot-OliverJ.R, R/polar_plot.R
 \name{polar_plot}
 \alias{polar_plot}
 \title{Generates a polar plot with elliptical confidence intervals}
 \usage{
+polar_plot(
+  x,
+  ci_level = 0.95,
+  n_breaks = 5,
+  component_index = NULL,
+  grid_angle_segments = 8,
+  radial_units = c("radians", "degrees", "period"),
+  clockwise = FALSE,
+  text_size = 3,
+  text_opacity = 0.5,
+  fill_colours,
+  ellipse_opacity = 0.3,
+  circle_linetype = "dotted",
+  start = c("right", "left", "top", "bottom"),
+  view = c("full", "zoom", "zoom_origin"),
+  overlay_parameter_info = FALSE,
+  quietly = TRUE,
+  show_component_labels = TRUE,
+  xlims,
+  ylims,
+  ...
+)
+
 polar_plot(
   x,
   ci_level = 0.95,
@@ -119,11 +142,21 @@ the polar plots have labels corresponding to their components.}
 }
 \value{
 Returns a \code{ggplot} object.
+
+Returns a \code{ggplot} object.
 }
 \description{
 Generates a polar plot with elliptical confidence intervals
+
+Generates a polar plot with elliptical confidence intervals
 }
 \examples{
+data(vitamind)
+model <- cglmm(
+  vit_d ~ X + amp_acro(time, group = "X", period = 12),
+  data = vitamind
+)
+polar_plot(model)
 data(vitamind)
 model <- cglmm(
   vit_d ~ X + amp_acro(time, group = "X", period = 12),
