### 2024-01-09 #603 @ropensci-review-bot
Approved! Thanks @RWParsons for submitting and @sachsmc, @jcavieresg for your reviews! :grin:
