### 2024-01-09 ropensci/GLMMcosinor#4 @RWParsons
@jcavieresg approved changes here: https://github.com/ropensci/software-review/issues/603#issuecomment-1882632517
