### 2024-01-09 ropensci/GLMMcosinor#2 @RWParsons
@sachsmc approved changes here: https://github.com/ropensci/software-review/issues/603#issuecomment-1881301247
